chore: update checkpoints, improve sync state notifications, create additional ui for sync progress indication, include coinjoin-related addresses for bloom filtering, add address registration test, refactor transaction data serialization

+++ DashSync/shared/Models/Managers/Chain Managers/DSChainManager+Protected.h
@@ -34,7 +34,7 @@ typedef NS_ENUM(uint16_t, DSChainNotificationType) {
 - (void)wipeMasternodeInfo;
 
 - (void)notifySyncStateChanged;
-- (void)notifyMasternodeSyncStateChange:(uint32_t)lastBlockHeihgt storedCount:(uintptr_t)storedCount;
+- (void)notifyMasternodeSyncStateChange:(uint32_t)lastBlockHeihgt storedCount:(uint32_t)storedCount;
 
 
 @end
